requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
opensearch-py
lancedb
python-multipart
PyJWT
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
mongomock==4.1.2
docling
pypdf
pdfplumber
PyPDF2
pymupdf
pdfminer.six
python-docx
python-dotenv
# Use llama-cpp-python binary wheels (pre-built, no compilation needed)
llama-cpp-python==0.2.76
google-generativeai
